backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# AI/ML Libraries
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
openai==1.3.0
langchain==0.0.340
chromadb==0.4.18
huggingface-hub==0.19.4
accelerate==0.25.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
django-oauth-toolkit==1.7.1
cryptography==41.0.8

# Monitoring & Performance
django-debug-toolbar==4.2.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# WebRTC & Media
aiortc==1.6.0
opencv-python==4.8.1.78
ffmpeg-python==0.2.0